1.20

+++ source/CapCalculators/AdvancedTrainingCalculator.cs
@@ -23,7 +23,7 @@ namespace jshepler.ngu.mods.CapCalculators
             return L >= long.MaxValue ? long.MaxValue : (long)L;
         }
 
-        protected override double GetProgressPerTick(long r, long L)
+        protected override double GetProgressPerTick(double r, double L)
         {
             return r / 50 * Math.Sqrt(p) * e / (b * L);
         }
